# specs/017-better-auth-integration/contracts/auth-api-contract.yaml
openapi: 3.0.3
info:
  title: Better Auth Integration API
  description: API contract for authentication and user isolation
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
paths:
...
  /api/{user_id}/tasks:
    get:
      summary: Get user's tasks
      description: Retrieves all tasks belonging to the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The authenticated user's ID
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: Bearer token
      responses:
        '200':
          description: Tasks retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TodoItem'
        '401':
          description: Unauthorized - invalid or missing token
        '403':
          description: Forbidden - user trying to access another user's data
    post:
      summary: Create a new task for user
      description: Creates a new task for the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The authenticated user's ID
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: Bearer token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoItem'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/TodoItem'
        '401':
          description: Unauthorized - invalid or missing token
        '403':
          description: Forbidden - user trying to access another user's data
  /api/{user_id}/tasks/{task_id}:
    get:
      summary: Get a specific task
      description: Retrieves a specific task for the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The authenticated user's ID
        - name: task_id
          in: path
          required: true
          schema:
            type: string
          description: The task ID
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: Bearer token
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/TodoItem'
        '401':
          description: Unauthorized - invalid or missing token
        '403':
          description: Forbidden - user trying to access another user's data
        '404':
          description: Task not found
    put:
      summary: Update a task
      description: Updates a specific task for the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The authenticated user's ID
        - name: task_id
          in: path
          required: true
          schema:
            type: string
          description: The task ID
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: Bearer token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoItem'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/TodoItem'
        '401':
          description: Unauthorized - invalid or missing token
        '403':
          description: Forbidden - user trying to access another user's data
        '404':
          description: Task not found
    delete:
      summary: Delete a task
      description: Deletes a specific task for the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The authenticated user's ID
        - name: task_id
          in: path
          required: true
          schema:
            type: string
          description: The task ID
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: Bearer token
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/TodoItem'
        '401':
          description: Unauthorized - invalid or missing token
        '403':
          description: Forbidden - user trying to access another user's data
        '404':
          description: Task not found
  /api/{user_id}/tasks/{task_id}/toggle:
    patch:
      summary: Toggle task completion status
      description: Toggles the completion status of a specific task for the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The authenticated user's ID
        - name: task_id
          in: path
          required: true
          schema:
            type: string
          description: The task ID
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: Bearer token
      responses:
        '200':
          description: Task completion status toggled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/TodoItem'
        '401':
          description: Unauthorized - invalid or missing token
        '403':
          description: Forbidden - user trying to access another user's data
        '404':
          description: Task not found
components:
  schemas:
    TodoItem:
      type: object
      required:
        - id
        - title
        - completed
        - user_id
      properties:
        id:
          type: string
        title:
          type: string
          maxLength: 255
        description:
          type: string
          nullable: true
          maxLength: 1000
        completed:
          type: boolean
        priority:
          type: string
          enum: [high, medium, low]
        category:
          type: string
          maxLength: 50
        user_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateTodoItem:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          nullable: true
          maxLength: 1000
        priority:
          type: string
          enum: [high, medium, low]
          default: medium
        category:
          type: string
          maxLength: 50
          default: personal
    UpdateTodoItem:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          nullable: true
          maxLength: 1000
        completed:
          type: boolean
        priority:
          type: string
          enum: [high, medium, low]
        category:
          type: string
          maxLength: 50